+++ weblate/addons/forms.py
@@ -4,9 +4,9 @@
 
 from __future__ import annotations
 
-import re
 from typing import TYPE_CHECKING
 
+import regex
 from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Field, Layout
 from django import forms
@@ -27,6 +27,7 @@ from weblate.utils.forms import (
     SortedSelectMultiple,
     WeblateServiceURLField,
 )
+from weblate.utils.regex import compile_regex
 from weblate.utils.render import validate_render, validate_render_translation
 from weblate.utils.validators import (
     DomainOrIPValidator,
@@ -290,6 +291,7 @@ class DiscoveryForm(BaseAddonForm):
                             "matches_created": created,
                             "matches_matched": matched,
                             "matches_deleted": deleted,
+                            "matches_errors": self.discovery.errors,
                             "matches_skipped": skipped,
                             "user": self.user,
                         },
@@ -345,8 +347,8 @@ class DiscoveryForm(BaseAddonForm):
         if "match" not in self.cleaned_data:
             return None
         try:
-            return re.compile(self.cleaned_data["match"])
-        except re.error:
+            return compile_regex(self.cleaned_data["match"])
+        except regex.error:
             return None
 
     def test_render(self, value):