+++ weblate/addons/tasks.py
@@ -18,15 +18,20 @@ from django.utils.timezone import now
 from lxml import html
 
 from weblate.addons.events import AddonEvent
-from weblate.addons.models import Addon, AddonActivityLog, handle_addon_event
+from weblate.addons.models import (
+    Addon,
+    AddonActivityLog,
+    handle_addon_event,
+    handle_daily_addon_event,
+)
 from weblate.lang.models import Language
 from weblate.trans.exceptions import FileParseError
 from weblate.trans.models import Change, Component, Project
 from weblate.utils.celery import app
 from weblate.utils.hash import calculate_checksum
 from weblate.utils.lock import WeblateLockTimeoutError
-from weblate.utils.requests import http_request
-from weblate.utils.validators import validate_asset_url, validate_filename
+from weblate.utils.requests import asset_request
+from weblate.utils.validators import validate_filename
 
 IGNORED_TAGS = {"script", "style"}
 
@@ -54,8 +59,7 @@ def cdn_parse_html(addon_id: int, component_id: int) -> None:
         filename = filename.strip()
         try:
             if filename.startswith(("http://", "https://")):
-                validate_asset_url(filename)
-                with http_request("get", filename) as handle:
+                with asset_request("get", filename) as handle:
                     content = handle.text
             else:
                 content = read_component_file(component, filename)
@@ -168,23 +172,13 @@ def language_consistency(
 
 @app.task(trail=False)
 def daily_addons(modulo: bool = True) -> None:
-    def daily_callback(
-        addon: Addon, component: Component, *, activity_log_id: int | None = None
-    ) -> None:
-        addon.addon.daily(component, activity_log_id=activity_log_id)
-
     today = timezone.now()
     addons = Addon.objects.filter(event__event=AddonEvent.EVENT_DAILY).prefetch_related(
         "component", "project"
     )
     if modulo:
         addons = addons.annotate(hourmod=F("id") % 24).filter(hourmod=today.hour)
-    handle_addon_event(
-        AddonEvent.EVENT_DAILY,
-        daily_callback,
-        addon_queryset=addons,
-        auto_scope=True,
-    )
+    handle_daily_addon_event(addons)
 
 
 def update_addon_activity_log(